Move editor toolbar to separate folder
