+++ packages/js/product-editor/src/components/iframe-editor/header-toolbar/header-toolbar.scss
