+++ packages/js/product-editor/src/components/iframe-editor/iframe-editor.tsx
@@ -27,7 +27,7 @@ import {
 import { BackButton } from './back-button';
 import { EditorCanvas } from './editor-canvas';
 import { EditorContext } from './context';
-import { HeaderToolbar } from './header-toolbar';
+import { HeaderToolbar } from './header-toolbar/header-toolbar';
 import { ResizableEditor } from './resizable-editor';
 import { SecondarySidebar } from './secondary-sidebar/secondary-sidebar';
 import { useEditorHistory } from './hooks/use-editor-history';
