+++ packages/js/product-editor/src/components/iframe-editor/style.scss
@@ -1,5 +1,5 @@
 @import './iframe-editor.scss';
-@import './header-toolbar.scss';
+@import './header-toolbar/header-toolbar.scss';
 @import './resize-handle.scss';
 @import './back-button.scss';
 @import './secondary-sidebar/inserter-sidebar.scss';
